+++ clang/lib/CodeGen/EHScopeStack.h
@@ -150,6 +150,8 @@ public:
     Cleanup(Cleanup &&) {}
     Cleanup() = default;
 
+    virtual bool isRedundantBeforeReturn() { return false; }
+
     /// Generation flags.
     class Flags {
       enum {